chd: cast before shifting in get_bigendian_uint32_t - signed overflow on ordinary data
    return (base[0] << 24) | (base[1] << 16) | (base[2] << 8) | base[3];

base[0] is a uint8_t. The usual arithmetic conversions promote it to int,
so shifting left by 24 overflows a signed int for any byte from 0x80 up,
which is undefined behaviour.

This is not an exotic-input bug. The function has 27 call sites in this
file covering header fields, map entries and metadata tags, and any CRC
or SHA-1 byte has a one-in-two chance of tripping it - so it fires while
parsing ordinary, well-formed CHD images, not only corrupt ones. The
neighbouring get_bigendian_uint64_t and get_bigendian_uint48 already cast
each byte for exactly this reason, so the 32-bit reader was an oversight
rather than a decision. get_bigendian_uint24 and get_bigendian_uint16
cannot overflow an int and are left alone.

The fix is behaviour-preserving. Verified over all 2^32 byte patterns
that the corrected reader returns exactly what the old one returned on a
two's-complement machine, comparing against the old expression evaluated
in a wider type so the comparison itself stays defined. Zero differences.

Found by fuzzing libchdr under ASan and UBSan through a driver covering
chd_open, the metadata walk and chd_read, seeded with a conformant
uncompressed CHD v5 image built to the layout in header_read. Byte-flip
mutation reaches it within a few hundred iterations; three separate
reproducers all landed on this line. After the fix the same fuzzer seed
runs 2000 iterations with no findings, and the conformant seed still
opens and decodes hunk 0 identically.

That this went unnoticed reflects coverage rather than the bug being
marginal: a UBSan sweep only catches it if something parses a CHD during
the run.

+++ libretro-common/formats/libchdr/libchdr_chd.c
@@ -585,7 +585,15 @@ static INLINE void put_bigendian_uint48(uint8_t *base, uint64_t value)
 
 static INLINE uint32_t get_bigendian_uint32_t(const uint8_t *base)
 {
-	return (base[0] << 24) | (base[1] << 16) | (base[2] << 8) | base[3];
+	/* Each byte is cast before shifting. base[0] is a uint8_t, which the
+	 * usual arithmetic conversions promote to int, so base[0] << 24 is a
+	 * signed shift that overflows for any byte from 0x80 up - undefined
+	 * behaviour, and it fires on ordinary data rather than exotic input:
+	 * a CRC or a SHA-1 byte has a one-in-two chance of tripping it. The
+	 * uint64 and uint48 readers above already cast for exactly this
+	 * reason, so this one was an oversight rather than a decision. */
+	return ((uint32_t)base[0] << 24) | ((uint32_t)base[1] << 16) |
+	       ((uint32_t)base[2] <<  8) |  (uint32_t)base[3];
 }
 
 /*-------------------------------------------------
